Revamp build system, site interface; add HTML metadata.
This commit changes a lot.

I changed the site's CSS slightly to add a shadow to cards. I think, this
way, they're a bit easier to discern from the background image.

I added metadata to the website, which should make it appear nicer when the
site is linked to on, for example, Discord. It is a bit of a mess though,
and the implementation needs some serious work. Metadata descriptions are
dynamically generated for most titles and markdown descriptions, however
raw HTML documents used in the site cannot utilise automatic generation at
this moment. I just hardcoded HTML documents' description content for the
time being.

I reorganised the build system a little bit, and separated the registration
process from the main build script. This should make it much nicer to work
with as the site grows.

I added a proper submodules definition in Git, although I'm not sure how
this will work out on GitHub until I push this commit. My original
directory structure for this repository was that public/ is the original
repo pushed to the kawapure.github.io repository. Git handles this just
fine locally, but the GitHub website just has a null submodule folder link
displayed.

Signed-off-by: Isabella <165211464+kawapure@users.noreply.github.com>